+++ modules/saxs/include/ChiScoreLog.h
@@ -20,18 +20,18 @@ IMPSAXS_BEGIN_NAMESPACE
 */
 class IMPSAXSEXPORT ChiScoreLog {
  public:
-  Float compute_score(const Profile* exp_profile, const Profile* model_profile,
-                      bool use_offset = false) const;
+  double compute_score(const Profile* exp_profile, const Profile* model_profile,
+                       bool use_offset = false) const;
 
-  Float compute_score(const Profile* exp_profile, const Profile* model_profile,
-                      Float min_q, Float max_q) const;
+  double compute_score(const Profile* exp_profile, const Profile* model_profile,
+                       double min_q, double max_q) const;
 
-  Float compute_scale_factor(const Profile* exp_profile,
-                             const Profile* model_profile,
-                             Float offset = 0.0) const;
+  double compute_scale_factor(const Profile* exp_profile,
+                              const Profile* model_profile,
+                              double offset = 0.0) const;
 
-  Float compute_offset(const Profile* exp_profile,
-                       const Profile* model_profile) const {
+  double compute_offset(const Profile* exp_profile,
+                        const Profile* model_profile) const {
     // not implemented as no straightforward solution to the equations
     IMP_UNUSED(exp_profile);
     IMP_UNUSED(model_profile);
